Simplify EnumConfigConvert derivation
